% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_with_laz.R
\name{convertLAZ}
\alias{convertLAZ}
\title{Converts LAZ to LAS + Index}
\usage{
convertLAZ(lazfile, outdir = "data/las", filter = "-keep_class 2 9")
}
\arguments{
\item{lazfile}{- LAZ file name}

\item{outdir}{= "data/las" - output directory,}

\item{filter}{= "-keep_class 2 9" - optional filter}
}
\description{
Converts LAZ to LAS + Index
}
