% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{Create Map}
\usage{
create_map(data,color_range = c("#FFFFFF","#ff2323"),color_option =
  "D",map_name = NULL, save_jpg = list(T, width = 80, height = 64),
  cathegorical = F, limits = NULL)
}
\arguments{
\item{data}{Data frame or Matrix with 3 columns: data, lon, lat in that order.}

\item{color_range}{Hexa decimal or colour names to describe the range in
which the color values will be ploted}

\item{color_option}{Color option for package \link{viridis} in case of
cathegorical values}

\item{map_name}{Name of the map used for saving and as a map tittle}

\item{save_jpg}{list of width and height values for jpeg file}

\item{cathegorical}{Boolean to define if the values are continuous or
cathegories}

\item{limits}{If defined, sets up the upper and lower bounds to the values
plotted in the map.}
}
\description{
Wrapper function to automate the process of creating maps and
  save them in jpg.
}
\examples{
create_map(map_change, save_jpg = list(T, width = 80, height = 64), map_name = "climate scenario 1",
          cathegorical = F, color_range = c("#FF0F47", "#FFFFFF", "#59BF34DB"),
          limits = c(-75,75))
}
\author{
Marcos Alves \email{mppalves@gmail.com}
}
