% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_gugik_data.R
\name{download_gugik_dataset_for_municipality}
\alias{download_gugik_dataset_for_municipality}
\title{download_gugik_dataset_for_municipality() downloads boundaries of villages within municipality,
downloads the dataset from GUGiK and saves it to file}
\usage{
download_gugik_dataset_for_municipality(municipiality, output_dir = "data")
}
\arguments{
\item{municipiality}{municipality name}

\item{output_dir}{default "data"}
}
\description{
download_gugik_dataset_for_municipality() downloads boundaries of villages within municipality,
downloads the dataset from GUGiK and saves it to file
}
