% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_input_gams.R
\name{export_weights_gams}
\alias{export_weights_gams}
\title{Export neural nets weights and biases to .csv}
\usage{
export_weights_gams(model, inputs_vec, data_set)
}
\arguments{
\item{model}{keras .h5 saved model.}

\item{inputs_vec}{vector with input column names}

\item{data_set}{Name used to identify the dataset being used}
}
\description{
Save and export model weights and biases for use in GAMS
  reconstruction of neural networks.
}
\examples{
## Manually extracted column names
inputs = c("lsu","lon","lat","tem","pre","rad", "soil")

## Suggested way to extract the inputs from the dataset
inputs = c(colnames(df.harvest)[-ncol(df.harvest)])
export_weights_gams(model, inputs, "harvest")
}
\author{
Marcos Alves \email{mppalves@gmail.com}
}
