% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_gugik_data.R
\name{format_file_name}
\alias{format_file_name}
\title{municipiality_file_name() creates and returns file name for municipality}
\usage{
format_file_name(name, dir, prefix, suffix, extension)
}
\arguments{
\item{name}{Municipality name}

\item{dir}{Output directory}

\item{prefix}{optional prefix to file name, like prefix_filename.ext}

\item{suffix}{optional suffix to file name, like filename_suffix.ext}

\item{extension}{file extension, default .rds}
}
\value{
file_name Name of file to read / write data
}
\description{
municipiality_file_name() creates and returns file name for municipality
}
\examples{
format_file_name("Oborniki Śląskie", "data")
format_file_name("Oborniki Śląskie", "data", extension = ".csv")

}
