% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_gugik_data.R
\name{get_gugik_data}
\alias{get_gugik_data}
\title{Downloads gugiik data to specific folder using wget}
\usage{
get_gugik_data(url, sha, output_dir)
}
\arguments{
\item{url}{url from GUGiK}

\item{sha}{optionally sha1 of the file}

\item{output_dir}{local folder to store data}
}
\description{
Downloads gugiik data to specific folder using wget
}
