% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstools.R
\docType{package}
\name{GSTools}
\alias{GSTools}
\title{Grass Tools}
\description{
Group of functions and warpper functions to pre-process and analyse data from LPJmL
}
\details{
List of functions in this package:

\code{\link{create_map}}

\code{\link{export2gams}}

\code{\link{generate_test_data}}

\code{\link{optimize_lsu}}

\code{\link{ploting_individual_cells}}

\code{\link{reconstruct_dataset}}

\code{\link{run_analysis}}
}
\author{
Marcos Alves \code{\email{mppalves@gmail.com}}
}
