% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_gugik_data.R
\name{link_gugik_data_from_archive}
\alias{link_gugik_data_from_archive}
\title{link_gugik_data_from_archive() creates a symbolic link to file in archive}
\usage{
link_gugik_data_from_archive(url, archive_dir, output_dir)
}
\arguments{
\item{url}{url from GUGiK}

\item{archive_dir}{folder path with archive data}

\item{output_dir}{local folder to store data}
}
\description{
link_gugik_data_from_archive() creates a symbolic link to file in archive
}
