% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_gugik_data.R
\name{municipiality_file_name}
\alias{municipiality_file_name}
\title{municipiality_file_name() creates and returns file name for municipality}
\usage{
municipiality_file_name(municipiality, dir, extension)
}
\arguments{
\item{municipiality}{Municipality name}

\item{dir}{Output directory}

\item{extension}{file extension, default .rds}
}
\value{
file_name Name of file to read / write data
}
\description{
municipiality_file_name() creates and returns file name for municipality
}
\examples{
municipiality_file_name("Oborniki Śląskie", "data")
municipiality_file_name("Oborniki Śląskie", "data", extension = ".csv")

}
