% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_lsu.R
\name{optimize_lsu}
\alias{optimize_lsu}
\title{Optmize LSUs}
\usage{
Optimize_LSU(test_data, model, min_Lsu, max_Lsu, col_means,
  col_stddevs, range)
}
\arguments{
\item{test_data}{Generated test data object, for mode details see
\link{generate_test_data}}

\item{model}{.h5 saved keras model}

\item{min_Lsu}{lower boundary for constrained optmization}

\item{max_Lsu}{upper boundary for constrained optmization}

\item{col_means}{Center attributes from scaled training data of the model
used in the parameter model.
Example: train_data <- scale(train_data)
col_means_train_harvest <- attr(train_data, "scaled:center")}

\item{col_stddevs}{Scale attributes from scaled training data of the model
used in the parameter model.}

\item{range}{Cell interval to be optmized}
}
\value{
Vector of optimized LSUs per grid cell
}
\description{
Optimize LSU for each cell using a neural net model.
}
\examples{
Optimize_LSU(test_data, ML_model_harvest,min_Lsu = -2, max_Lsu = 2,
                       col_means_train_harvest,col_stddevs_train_harvest,
                       range = 28415:28416)
}
\author{
Marcos Alves \email{mppalves@gmail.com}
}
