% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting_individual_cells.R
\name{ploting_individual_cells}
\alias{ploting_individual_cells}
\title{Plotting individual cells behaviour in 3D}
\usage{
ploting_individual_cells(test_data,cell,Lsu_range, climate_variable, climate_variable_range, ML_model)
}
\arguments{
\item{test_data}{Output from the function test_data containing neat data for
machine learning model tests.}

\item{cell}{Cell number from 1 to 67420.}

\item{climate_variable}{One of the following: "temperature_mean", "precipitation_mean", "radiation_mean"}

\item{climate_variable_range}{Normalized ploting range ex: from -1.5 to 2.5}

\item{ML_model}{Trained machine learning model saved in hd5}

\item{lsu_range}{Normalized ploting range ex: from -1.5 to 2.5}
}
\description{
Function to plot selected cells response to changes in LSU and
 climate variables
}
\examples{
ploting_individual_cells(test_data=test_data, cell = 28415,Lsu_range =c(-1.5,1.5) ,climate_variable = "precipitation_mean",
                                  climate_variable_range = c(-1.5,1.5), ML_model_harvest)
}
\author{
Marcos Alves \email{mppalves@gmail.com}
}
