% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_dataset.R
\name{reconstruct_dataset}
\alias{reconstruct_dataset}
\title{Reconstruct data set}
\usage{
reconstruct_dataset(test_data_unscaled, harvest_predicted)
}
\arguments{
\item{test_data_unscaled}{generate test data object with scale parameter set to FALSE (scale = F). For mode details see \link{generate_test_data}.}

\item{harvest_predicted}{data frame with the predicted harvest value}
}
\description{
Function to reconstruct in spread format the table with the values predicted by the model
}
\examples{
reconstruct_dataset(test_data_unscaled, harvest_predicted)
}
\author{
Marcos Alves \email{mpplaves@gmail.com}
}
