% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{Run_simulation}
\alias{Run_simulation}
\alias{as.function}
\title{Run simulation}
\usage{
Run_simulation(func, w, x, input_data, title, cor ="red", comment = NULL)
}
\arguments{
\item{w}{LSU values}

\item{x}{secondary parameter (preciptation, radiation, temperature or mean
harvest for example)}

\item{func}{R formula to be evaluated}

\item{input_data}{data frame with three columns: x variable, LSU, y output}

\item{title}{String used to name the plots and the exported files}

\item{cor}{hexadecimal or color name used to color details in plots}

\item{comment}{Comment useful to copy the formula being analysed}
}
\description{
Function to take as input a testing formula, LSUs and predictors
 and output staticts and plots that describe how good the formula is
 simmulating the original data.
}
\examples{

Run_simulation(func = y ~ 8.76366*exp(-0.0163042*x + (-1.56578 + 0.0103395*x)*w)*x*w,
              w=input_data$w,
              x=input_data$x,
              input_data = input_data,
              title = "Formula Evaluation Harvest, Yield mean",
              cor = "#217A00",
              comment = "y ~ 8.76366*exp(-0.0163042*x + (-1.56578 + 0.0103395*x)*w)*x*w")
}
\author{
Marcos Alves \email{mppalves@gmail.com}
}
