% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstools.R
\docType{package}
\name{GSTools}
\alias{GSTools}
\alias{GSTools-package}
\title{Grass Tools}
\description{
Group of functions and warpper functions to pre-process and analyse data from LPJmL
}
\details{
List of functions in this package:

\code{\link{create_map}}

\code{\link{export_weights_gams}}

\code{\link{generate_test_data}}

\code{\link{Optimize_LSU}}

\code{\link{ploting_individual_cells}}

\code{\link{reconstruct_dataset}}

\code{\link{Run_analysis}}

\code{\link{Run_simulation}}
}
\author{
Marcos Alves \code{\email{mppalves@gmail.com}}
}
